requirements.txt
ollama
PyQt5
SpeechRecognition
pyaudio
edge-tts
pygame
google-generativeai
textblob
numpy
piper-tts
youtube-search-python
duckduckgo-search
groq
